requirements.txt
altair==4.1.0
argcomplete==1.10.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.8.0
bleach==3.2.1
blinker==1.4
blis==0.4.1
boto3==1.15.18
botocore==1.18.18
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Counter==1.0.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
enum-compat==0.0.3
extract-msg==0.23.1
gitdb==4.0.5
GitPython==3.1.9
idna==2.10
IMAPClient==2.1.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.19.2
olefile==0.46
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pdfminer.six==20200517
pdfplumber==0.5.23
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pt-core-news-sm==2.3.0
ptyprocess==0.6.0
pyarrow==1.0.1
pycparser==2.20
pycryptodome==3.9.8
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
pyzmq==19.0.2
regex==2020.10.15
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.12.0
smmap==3.0.4
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
spacy-streamlit==0.1.0
SpeechRecognition==3.8.1
srsly==1.0.2
streamlit==0.69.2
terminado==0.9.1
testpath==0.4.4
textract==1.6.3
thinc==7.4.1
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
tzlocal==1.5.1
urllib3==1.25.11
validators==0.18.1
Wand==0.6.3
wasabi==0.8.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
xlrd==1.2.0
XlsxWriter==1.3.7
